WPF Graph showing valid data. More UI update needed.
