+++ tools/Profiler/NanoProfiler/NanoProfiler.csproj
@@ -29,4 +29,10 @@
     <Folder Include="Views\Controls\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Update="Views\GraphViewWinForm.xaml.cs">
+      <SubType>Code</SubType>
+    </Compile>
+  </ItemGroup>
+
 </Project>
