+++ tools/Profiler/NanoProfiler/Views/GraphViewWinForm.xaml
@@ -1,0 +1,142 @@
+﻿<Window x:Class="nanoFramework.Tools.NanoProfiler.Views.GraphViewWinForm"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:gridViewer="clr-namespace:WinForms.CLRProfiler"
+        xmlns:local="clr-namespace:nanoFramework.Tools.NanoProfiler.Views"
+        xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf" 
+        xmlns:clrprofiler="clr-namespace:WinForms.CLRProfiler"
+        mc:Ignorable="d"
+        Title="GraphView" MinHeight="800" MinWidth="1200">
+
+    <Window.ContextMenu>
+        <ContextMenu>
+            <MenuItem Header="Show Who Allocated" 
+               Click="MenuItem_Click"/>
+            <MenuItem Header="Export Data to File.." />
+        </ContextMenu>
+    </Window.ContextMenu>
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="100" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="*" />
+                <ColumnDefinition Width="*" />
+            </Grid.ColumnDefinitions>
+            <StackPanel Margin="10,0,0,0" VerticalAlignment="Center" Orientation="Horizontal">
+                <Label Width="163" Height="30" HorizontalAlignment="Left"
+                       Content="Scale:" FontSize="16" />
+                <ComboBox   Name="VerticalScaleOption" Width="120" Margin="20,0,0,0"
+                            SelectedItem="{Binding ScaleSelectedValue, Mode=TwoWay}" 
+                            ItemsSource="{Binding ScaleList}">
+
+                </ComboBox>
+            </StackPanel>
+
+            <StackPanel
+            Grid.Column="1"
+            Margin="10,0,0,0"
+            VerticalAlignment="Center"
+            Orientation="Horizontal">
+                <Label
+                Width="163"
+                Height="30"
+                HorizontalAlignment="Left"
+                Content="Detail:"
+                FontSize="16" />
+                <ComboBox   Name="HorizontalScaleOption" Width="120" Margin="20,0,0,0" HorizontalAlignment="Left"
+                            SelectedValue="{Binding DetailsSelectedValue, Mode=TwoWay}" 
+                            ItemsSource="{Binding DetailsList}">
+
+                </ComboBox>
+            </StackPanel>
+        </Grid>
+
+        <!--<Canvas x:Name="graphPanel" >
+            <Canvas Width="508" Height="209" />
+        </Canvas>-->
+        <Grid Grid.Row="1">
+
+                <WindowsFormsHost x:Name="windowsFormsHost" Visibility="Collapsed" 
+                                   >
+
+                </WindowsFormsHost>
+
+                <!--<WindowsFormsHost x:Name="windowsFormsHost" Grid.Column="0">
+                    <StaticResourceExtension ResourceKey="gridViewerInstance" />
+                </WindowsFormsHost>-->
+
+                <!--<WindowsFormsHost x:Name="windowsFormsHost" Visibility="Visible"   >
+                    -->
+                <!--<clrprofiler:GraphViewForm />-->
+                <!--
+                </WindowsFormsHost>-->
+
+
+
+                <!--Live CHarts / Pie Chart-->
+            <!--<lvc:CartesianChart Margin="20" >
+                <lvc:CartesianChart.Series >
+                    <lvc:ColumnSeries Title="Value"  Values="{Binding GraphValues}" DataLabels="True"
+                                          Configuration="{Binding GraphConfiguration}" StrokeThickness="1"/>
+                </lvc:CartesianChart.Series>
+                <lvc:CartesianChart.AxisX  >
+                    <lvc:Axis Foreground="Black" >
+                        <lvc:Axis.Separator>
+                            <lvc:Separator Step="1" />
+                        </lvc:Axis.Separator>
+                    </lvc:Axis>
+                </lvc:CartesianChart.AxisX>
+                <lvc:CartesianChart.AxisY  >
+                    <lvc:Axis Foreground="Black" MaxValue="100" MinValue="0" >
+                        <lvc:Axis.Separator>
+                            <lvc:Separator Step="10" />
+                        </lvc:Axis.Separator>
+                    </lvc:Axis>
+                </lvc:CartesianChart.AxisY>
+            </lvc:CartesianChart>-->
+
+            <!--<lvc:PieChart LegendLocation="Bottom" Hoverable="False" DataTooltip="{x:Null}">
+                <lvc:PieChart.Series>
+                    
+                    <lvc:PieSeries Title="1" Values="{Binding GraphValues}"
+                                   
+                                   Configuration="{Binding GraphConfiguration}"
+                                   DataLabels="True"
+                               LabelPoint="{Binding PointLabel}"/>
+                    --><!--<lvc:PieSeries Title="2" Values="4" DataLabels="True" 
+                               LabelPoint="{Binding PointLabel}"/>
+                    <lvc:PieSeries Title="3" Values="6" DataLabels="True" 
+                               LabelPoint="{Binding PointLabel}"/>
+                    <lvc:PieSeries Title="4" Values="2" DataLabels="True" 
+                               LabelPoint="{Binding PointLabel}"/>--><!--
+                </lvc:PieChart.Series>
+            </lvc:PieChart>-->
+
+
+            <!--<Grid>
+                <lvc:PieChart x:Name="myPieChart" />
+            </Grid>-->
+            <!--<Grid>
+                <lvc:PieChart Series="{Binding PieSeriesCollection}" 
+                              LegendLocation="Bottom" Hoverable="True" 
+                              
+                              
+                              />
+            </Grid>-->
+            <!--<lvc:PieSeries Title="Apples" 
+               Values="{Binding FruitPieChartModel.AppleDataSeries}" 
+               DataLabels="True"
+               LabelPoint="{Binding PointLabel}"/>-->
+
+
+        </Grid>
+
+    </Grid>
+</Window>
